+++ src/components/loadingPage.jsx
@@ -1,11 +1,11 @@
+import { Loader } from "lucide-react";
 import React from "react";
-import { ClipLoader } from "react-spinners";
 
 const LoadingPage = ({message}) => {
   return (
     <div className="min-h-screen bg-[#151518]">
       <div className="flex justify-center items-center min-h-screen">
-        <ClipLoader color="#3498db" loading={true} size={30} />
+        <Loader color="#3498db" loading={true} className="animate-spin" size={30} />
         <div className="text-white ml-4">{message}</div>
       </div>
     </div>
